requirements.txt
biopython==1.78
gensim==4.3.1
numpy==1.25.0
obonet==0.2.3
openpyxl==3.0.10
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.11.1
torch==2.0.1
torch-geometric==2.3.1
tqdm==4.65.0